Simplified argument handling for workers.
Instead of relying on an pseudo keyword arguments Hash the worker now requires
manually/explicitly defined constructor methods as well as requiring the mapper
to actually pass the right arguments and in the right order.

This change means that Ruby itself can validate whether or not the required
arguments are put in place. It's also a much more explicit way of doing it
which should make reading/understanding the code easier.