Extra error data from workers to the daemon.
Worker instances now have the ability to pass extra data back to the daemon
upon encountering an error. Using this setup you can pass worker input back
without having to re-raise custom error classes or having to pass in error
handling into the worker (which should be handled by the daemon instead).

Due to this being incompatible with existing code the version will be bumped to
2.0.0 instead of 1.1.0.